requirements.txt
asttokens==3.0.0
Authlib==1.3.1
bcrypt==4.1.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQL==1.6.0
Flask-OAuthlib==0.9.6
Flask-WTF==1.2.1
fonttools==4.57.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==2.1.0
ordered-set==4.1.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32==310
pyzmq==26.3.0
requests==2.32.4
requests-oauthlib==1.1.0
rich==13.9.4
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
